[ALPS03301566] Audio: Fix build warnings
[Detail]
Remove integer type define.

MTK-Commit-Id: ffadeb5a5e542276fb878e0e20bb9eccd4b1acdf

CR-Id: ALPS03301566
Feature: Audio output

+++ sound/soc/mediatek/common_int/mtk-soc-pcm-btcvsd.h
@@ -204,8 +204,8 @@ struct btcvsd_tx_buffer_info {
 };
 
 struct time_buffer_info {
-	uint64 data_count_equi_time;
-	uint64 timestamp_us;
+	unsigned long long data_count_equi_time;
+	unsigned long long timestamp_us;
 };
 
 struct cvsd_memblock {
